Address code review findings
- Fix event monitor leak: add .onDisappear cleanup to ShortcutRecorderView
- Add isLoading flag to suppress redundant save() during load()
- Extract shared NSOpenPanel helper to eliminate DRY violation
- Add reserved shortcut denylist to prevent overriding system shortcuts
- Pass through unrecognized keys instead of swallowing all events
- Remove unused import Combine
